requirements.txt
Django==4.1
djangorestframework==3.13.1
Pillow==9.2.0
django-environ==0.9.0

django-simple-history==3.1.1

django-cors-headers==3.13
psycopg2==2.9.4